refactor(forge): split GitHub GraphQL transport into github_graphql.go (#1025)
Pure same-package code motion: move graphqlEndpoint, graphqlURL, the
graphqlError/graphqlResponse envelope types, and graphqlDo (with the
"Raw GraphQL client for Projects v2" section comment that documents them)
out of github.go into a new sibling github_graphql.go. No behaviour
change; go doc -u -all is byte-identical.

The section comment's stale positional reference "github.go:610" (which
pointed at where the go-github auth transport is built) was repointed to
name the identifier newGitHub instead, since a line number does not
survive the move.
